use `test_db_prefix_scan` for all 4 database types
